[refactor] Share common code between RestoreHandler and SystemImportHandler
